for #2556, refactor package of parse test
